refactor: 공통 유효성 결과값 재사용 객체 생성
