+++ hooks/src/lib/constants/validation.ts
@@ -4,3 +4,8 @@ export const ERROR_MESSAGE = {
   INVALID_MONTH: "유효한 월(MM)의 범위를 입력해주세요.(01~12)",
   INVALID_YEAR: "유효한 연도(YY)를 입력해주세요.",
 };
+
+export const defaultValidationValue = {
+  isError: false,
+  errorMessage: null,
+};
